refactor(eventplatformreceiver): migrate comp/forwarder/eventplatformreceiver to V2 component architecture
