--- comp/forwarder/eventplatformreceiver/component_mock.go
@@ -1,15 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2024-present Datadog, Inc.
-
-//go:build test
-
-package eventplatformreceiver
-
-// team: agent-log-pipelines
-
-// Mock is the component type.
-type Mock interface {
-	Component
-}
