+++ comp/forwarder/eventplatformreceiver/def/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "def",
+    srcs = ["component.go"],
+    importpath = "github.com/DataDog/datadog-agent/comp/forwarder/eventplatformreceiver/def",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/logs/diagnostic",
+        "//pkg/logs/message",
+    ],
+)
