+++ comp/forwarder/eventplatformreceiver/def/component.go
