+++ comp/forwarder/eventplatformreceiver/fx/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "fx",
+    srcs = ["fx.go"],
+    importpath = "github.com/DataDog/datadog-agent/comp/forwarder/eventplatformreceiver/fx",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//comp/forwarder/eventplatformreceiver/impl",
+        "//pkg/util/fxutil",
+    ],
+)
