+++ comp/forwarder/eventplatformreceiver/impl/format.go
