+++ comp/forwarder/eventplatformreceiver/mock/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "mock",
+    srcs = ["mock.go"],
+    importpath = "github.com/DataDog/datadog-agent/comp/forwarder/eventplatformreceiver/mock",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//comp/forwarder/eventplatformreceiver/def",
+        "//pkg/logs/diagnostic",
+        "//pkg/logs/message",
+        "//pkg/util/fxutil",
+    ],
+)
